Move AuthApplication to AuthApplication.cs
