+++ tests/AuthEndpoints.MinimalApi.Tests/AuthApplication.cs
@@ -1,0 +1,28 @@
+﻿using AuthEndpoints.Demo.Data;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Storage;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Hosting;
+//using Microsoft.VisualStudio.TestPlatform.TestHost;
+
+namespace AuthEndpoints.MinimalApi.Tests;
+
+class AuthApplication : WebApplicationFactory<Program>
+{
+    protected override IHost CreateHost(IHostBuilder builder)
+    {
+        var root = new InMemoryDatabaseRoot();
+
+        builder.ConfigureServices(services =>
+        {
+            services.RemoveAll(typeof(DbContextOptions<MyDbContext>));
+
+            services.AddDbContext<MyDbContext>(options =>
+                options.UseInMemoryDatabase("Testing", root));
+        });
+
+        return base.CreateHost(builder);
+    }
+}
